+++ cnccode/controller_classes/CTManagementReports.inc.php
@@ -361,6 +361,24 @@ class CTManagementReports extends CTCNC
 
     } // end function sepndBycategory
 
+    /**
+     * @param $year
+     * @param null|integer $customerId
+     * @param null|integer $sector
+     * @param null| $pcs
+     */
+    function salesData(
+        $year,
+        $customerId = null,
+        $sector = null,
+        $pcs = null
+    )
+    {
+
+        $results = $this->buManagementReports->getSalesByCustomer($_REQUEST['customerID'], $_REQUEST['year']);
+
+    }
+
     function SalesByCustomer()
     {
         $this->setMethodName('SalesByCustomer');
@@ -382,7 +400,7 @@ class CTManagementReports extends CTCNC
                 )
             );
 
-        $results = $this->buManagementReports->getSalesByCustomer($_REQUEST['customerID'], $_REQUEST['year']);
+//        $results = $this->buManagementReports->getSalesByCustomer($_REQUEST['customerID'], $_REQUEST['year']);
 
         if ($_REQUEST['customerID']) {
             $dbeCustomer = new DBECustomer($this);
@@ -390,57 +408,56 @@ class CTManagementReports extends CTCNC
             $customerName = $dbeCustomer->getValue('name');
         }
 
-        $this->template->set_block('ManagementReportsSalesCustomer', 'resultsBlock', 'results');
+//        $this->template->set_block('ManagementReportsSalesCustomer', 'resultsBlock', 'results');
 
-        $minValue = 99999;
-        $maxValue = 0;
 
-        while ($row = $results->fetch_object()) {
-
-            $this->template->set_var(
-                array(
-                    'customer' => Controller::htmlDisplayText($row->customer),
-                    'sector' => Controller::htmlDisplayText($row->sector),
-                    'noOfPCs' => Controller::htmlDisplayText($row->noOfPCs),
-                    'noOfServers' => Controller::htmlDisplayText($row->noOfServers),
-                    'salesMonth1' => Controller::formatNumber($row->salesMonth1, 0),
-                    'profitMonth1' => Controller::formatNumber($row->profitMonth1, 0),
-                    'salesMonth2' => Controller::formatNumber($row->salesMonth2, 0),
-                    'profitMonth2' => Controller::formatNumber($row->profitMonth2, 0),
-                    'salesMonth3' => Controller::formatNumber($row->salesMonth3, 0),
-                    'profitMonth3' => Controller::formatNumber($row->profitMonth3, 0),
-                    'salesMonth4' => Controller::formatNumber($row->salesMonth4, 0),
-                    'profitMonth4' => Controller::formatNumber($row->profitMonth4, 0),
-                    'salesMonth5' => Controller::formatNumber($row->salesMonth5, 0),
-                    'profitMonth5' => Controller::formatNumber($row->profitMonth5, 0),
-                    'salesMonth6' => Controller::formatNumber($row->salesMonth6, 0),
-                    'profitMonth6' => Controller::formatNumber($row->profitMonth6, 0),
-                    'salesMonth7' => Controller::formatNumber($row->salesMonth7, 0),
-                    'profitMonth7' => Controller::formatNumber($row->profitMonth7, 0),
-                    'salesMonth8' => Controller::formatNumber($row->salesMonth8, 0),
-                    'profitMonth8' => Controller::formatNumber($row->profitMonth8, 0),
-                    'salesMonth9' => Controller::formatNumber($row->salesMonth9, 0),
-                    'profitMonth9' => Controller::formatNumber($row->profitMonth9, 0),
-                    'salesMonth10' => Controller::formatNumber($row->salesMonth10, 0),
-                    'profitMonth10' => Controller::formatNumber($row->profitMonth10, 0),
-                    'salesMonth11' => Controller::formatNumber($row->salesMonth11, 0),
-                    'profitMonth11' => Controller::formatNumber($row->profitMonth11, 0),
-                    'salesMonth12' => Controller::formatNumber($row->salesMonth12, 0),
-                    'profitMonth12' => Controller::formatNumber($row->profitMonth12, 0),
-                    'totalSales' => Controller::formatNumber($row->salesTotal, 0),
-                    'totalProfit' => Controller::formatNumber($row->profitTotal, 0)
-
-                )
-            );
-
-            $this->template->parse('results', 'resultsBlock', true);
-        }
+//        while ($row = $results->fetch_object()) {
+//
+//            $this->template->set_var(
+//                array(
+//                    'customer' => Controller::htmlDisplayText($row->customer),
+//                    'sector' => Controller::htmlDisplayText($row->sector),
+//                    'noOfPCs' => Controller::htmlDisplayText($row->noOfPCs),
+//                    'noOfServers' => Controller::htmlDisplayText($row->noOfServers),
+//                    'salesMonth1' => Controller::formatNumber($row->salesMonth1, 0),
+//                    'profitMonth1' => Controller::formatNumber($row->profitMonth1, 0),
+//                    'salesMonth2' => Controller::formatNumber($row->salesMonth2, 0),
+//                    'profitMonth2' => Controller::formatNumber($row->profitMonth2, 0),
+//                    'salesMonth3' => Controller::formatNumber($row->salesMonth3, 0),
+//                    'profitMonth3' => Controller::formatNumber($row->profitMonth3, 0),
+//                    'salesMonth4' => Controller::formatNumber($row->salesMonth4, 0),
+//                    'profitMonth4' => Controller::formatNumber($row->profitMonth4, 0),
+//                    'salesMonth5' => Controller::formatNumber($row->salesMonth5, 0),
+//                    'profitMonth5' => Controller::formatNumber($row->profitMonth5, 0),
+//                    'salesMonth6' => Controller::formatNumber($row->salesMonth6, 0),
+//                    'profitMonth6' => Controller::formatNumber($row->profitMonth6, 0),
+//                    'salesMonth7' => Controller::formatNumber($row->salesMonth7, 0),
+//                    'profitMonth7' => Controller::formatNumber($row->profitMonth7, 0),
+//                    'salesMonth8' => Controller::formatNumber($row->salesMonth8, 0),
+//                    'profitMonth8' => Controller::formatNumber($row->profitMonth8, 0),
+//                    'salesMonth9' => Controller::formatNumber($row->salesMonth9, 0),
+//                    'profitMonth9' => Controller::formatNumber($row->profitMonth9, 0),
+//                    'salesMonth10' => Controller::formatNumber($row->salesMonth10, 0),
+//                    'profitMonth10' => Controller::formatNumber($row->profitMonth10, 0),
+//                    'salesMonth11' => Controller::formatNumber($row->salesMonth11, 0),
+//                    'profitMonth11' => Controller::formatNumber($row->profitMonth11, 0),
+//                    'salesMonth12' => Controller::formatNumber($row->salesMonth12, 0),
+//                    'profitMonth12' => Controller::formatNumber($row->profitMonth12, 0),
+//                    'totalSales' => Controller::formatNumber($row->salesTotal, 0),
+//                    'totalProfit' => Controller::formatNumber($row->profitTotal, 0)
+//
+//                )
+//            );
+//
+//            $this->template->parse('results', 'resultsBlock', true);
+//        }
         $this->template->set_var(
             array(
-                'urlGenerateReport' => $urlGenerateReport,
                 'customerPopupURL' => $customerPopupURL,
                 'customerName' => $customerName,
-                'customerID' => $_REQUEST['customerID']
+                'customerID' => $_REQUEST['customerID'],
+                'sectorID' => $_REQUEST['sectorID'],
+                'noOfPcs' => $_REQUEST['noOfPcs']
             )
         );